CNN_Project/requirements.txt
# PyTorch and torchvision for CNN
torch>=2.0.0
torchvision>=0.15.0

# Kaggle API for dataset download
kagglehub>=0.2.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.8.0

# Data processing
pandas>=1.5.0
numpy>=1.23.0

# Progress bars
tqdm>=4.65.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Streamlit for web app
streamlit>=1.28.0

# Utilities
requests>=2.31.0
scikit-learn>=1.3.0
